move search_and_print into aur_utilities from utilities to get rid of a dependency loop
